app/src/main/java/galacticgadgets/galactic/android/bignerdranchdatabindingandmvvmstudy/SoundViewModel.kt
package galacticgadgets.galactic.android.bignerdranchdatabindingandmvvmstudy

import android.os.Debug
import android.util.Log
import androidx.databinding.BaseObservable
import androidx.databinding.Bindable

class SoundViewModel(private val beatBox: BeatBox): BaseObservable() {
    fun onButtonClicked() {
        sound?.let {
            beatBox.play(it)
        }
    }

    var sound: Sound? = null
        set(sound){
            field = sound
            notifyChange()
        }

    @get:Bindable
    val title: String?
        get() = sound?.name
}